SwiftWin32: further adopt static libraries on 5.7+
Swift 5.7.0 enables static libraries for non-standard library
cases.  Use this to extract the CoreGraphics interfaces out of the
primary library.
