+++ Package.swift
@@ -27,9 +27,12 @@ let SwiftWin32: Package =
           targets: [
             .target(name: "CoreAnimation",
                     path: "Sources/SwiftWin32/CoreAnimation"),
+            .target(name: "CoreGraphics",
+                    path: "Sources/SwiftWin32/CoreGraphics"),
             .target(name: "SwiftWin32",
                     dependencies: [
                       "CoreAnimation",
+                      "CoreGraphics",
                       .product(name: "Logging", package: "swift-log"),
                       .product(name: "OrderedCollections",
                                package: "swift-collections"),
@@ -37,7 +40,11 @@ let SwiftWin32: Package =
                       .product(name: "SwiftCOM", package: "swift-com"),
                     ],
                     path: "Sources/SwiftWin32",
-                    exclude: ["CoreAnimation", "CMakeLists.txt"],
+                    exclude: [
+                      "CoreAnimation",
+                      "CoreGraphics",
+                      "CMakeLists.txt"
+                    ],
                     swiftSettings: [
                       .enableExperimentalFeature("AccessLevelOnImport"),
                     ],
@@ -76,7 +83,7 @@ let SwiftWin32: Package =
                               ]),
             .target(name: "TestUtilities", path: "Tests/Utilities"),
             .testTarget(name: "AutoLayoutTests", dependencies: ["SwiftWin32"]),
-            .testTarget(name: "CoreGraphicsTests", dependencies: ["SwiftWin32"]),
+            .testTarget(name: "CoreGraphicsTests", dependencies: ["CoreGraphics"]),
             .testTarget(name: "SupportTests", dependencies: ["SwiftWin32"]),
             .testTarget(name: "UICoreTests",
                         dependencies: ["SwiftWin32", "TestUtilities"])
