+++ Sources/SwiftWin32/Animation and Haptics/Property-Based Animations/SpringTimingParameters.swift
@@ -1,6 +1,10 @@
 // Copyright © 2021 Saleem Abdulrasool <compnerd@compnerd.org>
 // SPDX-License-Identifier: BSD-3-Clause
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 /// The timing information for animations that mimics the behavior of a spring.
 public class SpringTimingParameters {
   // MARK - Initializing a Spring Timing Parameters Object
