+++ Sources/SwiftWin32/CoreGraphics/AffineTransform.swift
