+++ Sources/SwiftWin32/CoreGraphics/Point.swift
