+++ Sources/SwiftWin32/CoreGraphics/Size.swift
