+++ Sources/SwiftWin32/CoreGraphics/Vector.swift
