+++ Sources/SwiftWin32/Focus-Based Navigation/FocusItem.swift
@@ -1,6 +1,10 @@
 // Copyright © 2020 Saleem Abdulrasool <compnerd@compnerd.org>
 // SPDX-License-Identifier: BSD-3-Clause
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 /// An object that can become focused.
 public protocol FocusItem: FocusEnvironment {
   // MARK - Determining Focusability
