+++ Sources/SwiftWin32/Menus and Shortcuts/ContextMenuInteraction.swift
@@ -1,6 +1,10 @@
 // Copyright © 2020 Saleem Abdulrasool <compnerd@compnerd.org>
 // SPDX-License-Identifier: BSD-3-Clause
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 extension ContextMenuInteraction {
   /// Constants that describe the appearance of the menu.
   public enum Appearance: Int {
