+++ Sources/SwiftWin32/Pointer Interactions/PointerStyle.swift
@@ -1,6 +1,10 @@
 // Copyright © 2020 Saleem Abdulrasool <compnerd@compnerd.org>
 // SPDX-License-Identifier: BSD-3-Clause
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 extension PointerEffect {
   /// An effect that defines how to apply a tint to a view during a pointer
   /// interaction.
