+++ Sources/SwiftWin32/UI/SceneSizeRestrictions.swift
@@ -1,6 +1,10 @@
 // Copyright © 2020 Saleem Abdulrasool <compnerd@compnerd.org>
 // SPDX-License-Identifier: BSD-3-Clause
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 public class SceneSizeRestrictions {
   /// Setting the Size Restrictions
