+++ Sources/SwiftWin32/Views and Controls/Label.swift
@@ -3,6 +3,10 @@
 
 import WinSDK
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 private let SwiftLabelProc: SUBCLASSPROC = { (hWnd, uMsg, wParam, lParam, uIdSubclass, dwRefData) in
   let label: Label? = unsafeBitCast(dwRefData, to: AnyObject.self) as? Label
