+++ Sources/SwiftWin32/Views and Controls/Stepper.swift
@@ -3,6 +3,10 @@
 
 import WinSDK
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 // Notification Proxy
 
 // When the Window is created, the initial parent is cached.  This cache cannot
