+++ Sources/SwiftWin32/Windows and Screens/Window.swift
@@ -5,6 +5,10 @@ import WinSDK
 import class Foundation.NSNotification
 import class Foundation.NotificationCenter
 
+#if swift(>=5.7)
+import CoreGraphics
+#endif
+
 private let SwiftWindowProc: SUBCLASSPROC = { (hWnd, uMsg, wParam, lParam, uIdSubclass, dwRefData) in
   let window: Window? = unsafeBitCast(dwRefData, to: AnyObject.self) as? Window
