+++ Tests/UICoreTests/ViewTests.swift
@@ -2,6 +2,9 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 import XCTest
+#if swift(>=5.7)
+import CoreGraphics
+#endif
 @testable import SwiftWin32
 
 final class ViewTests: XCTestCase {
